Fix UBL-CR-528 warning
The warning was never triggered because it was checking for the line-level element cac:OrderLineReference/cbc:OrderReference.

OrderReference here is an aggregated element, just as the document-level OrderReference, so the check should be for cac:OrderReference. A document with cbc:OrderReference here would not pass XML Schema validation anyway.

This commit fixes that issue.