Fix zero comparison for -0.0 in BR-*-05 BR-*-06 BR-*-07 BR-*-09
As reported in https://github.com/SimplerInvoicing/validation/issues/22, the issue is that '<number> = 0' does not appear to return true when the value is -0.0, unless the element is cast with xs:decimal() first.

This is admittedly an edge case, resulting from creditnote conversions where all relevant values are multiplied by -1, but for consistency it would seem to be good to correctly evaluate -0.0 as well. This should also be consistent with other checks where xs:decimal() is used.