Fix issue with normalize-space for elements that can occur multiple times.
normalize-space() can only take a single string value. This works fine for xpath expressions that return a single node, but not for expressions that return a node list.

This results in a run-time XPath error when validating documents that have multiple nodes that match an expression within a context.

This commit adds string-join() to the test expression where multiple nodes may match the tests, according to the XML Schemais of UBL and CII.