Fix issue with normalize-space for elements that can occur multiple times.
normalize-space() can only take a single string value. This works fine for xpath expressions that return a single node, but not for expressions that return a node list.

This results in a run-time XPath error when validating documents that have multiple nodes that match an expression within a context.

This commit changes the test expressions for empty values to an 'every $ in satisfies' clause.

Note that most of these won't occur in a situation where the XML Schema is properly checked. However, because the result would be an Xpath error rather than a failed test, it's better to use this approach for all fields where we check emptiness by a normalize-space() != '' expression.