Fix variable name for $Credit_Payment_instructions
The use of the variable has a space after the name, while the definition does not. This results in the variable not being expanded when using the Schematron Skeleton implementation, and thus in an xlst file that can't be loaded.

It would probably make more sense to remove the space, but for consistency with the rest of the variables i've added it, rather than remove the one where the variable is used.