Remove rules CII-SR-465 and CII-SR-466.
It's perfectly reasonable to specify both person and department
names.

Rolling out those two rules will break a lot of formerly valid
invoices and cause a massive negative impact on electronic invoicing.

As mentioned by oriol in issue #448 (*not* #488 referenced in the
commit message for bf3a7d8d7137bec0c90f35d0fad42e30128fa2fd), the
problem with the cardinality for BT-41 and BT-56 isn't something
to be fixed in the schematron rules.

This reverts commit bf3a7d8d7137bec0c90f35d0fad42e30128fa2fd.